Levels weren't properly lit if OpenGL style or colored lighting were enabled when first playing a Descent 1 level after program launches, or when switching from playing a Descent 1 to a Descent 2 level or vice versa. Added options to enable built-in headlights and suppress energy consumption by headlights.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@927 01821946-851b-0410-b7f4-f1477c70e57a

+++ main/particles.c
@@ -846,7 +846,7 @@ else {
 #if 1
 	pc.alpha = (pc.alpha - 0.005f) * decay + 0.005f;
 	pc.alpha *= pc.alpha;
-	pc.alpha = pow (pc.alpha, 1.0 / 3.0);
+	pc.alpha = (float) pow (pc.alpha, 1.0 / 3.0);
 #else
 	pc.alpha = (float) cos ((double) (sqr (1 - decay) * Pi)) / 2 + 0.5f;
 #endif
